build(check): keep drift detection out of the generated artifact
Review on PR #1732 (HEAD 459a45ff, chatgpt-codex-connector): delegating
build:check's WHOLE enforcement (build, tracked-diff, untracked-file
check) into a single generated scripts/build-check.mjs meant a
committed copy that drifted from its source -- by accident or
tampering -- could still carry the idd-generated-from banner and exit
early, hiding its own drift with nothing outside it to catch that;
lint:minimum would return success without ever comparing it against
src/scripts/build-check.mts.

Move the build-then-tracked-diff comparison back into package.json's
own (non-generated) build:check line, restoring the original
pnpm-run-build-then-git-diff composition -- portable as-is, since it
uses only && (which cmd.exe on Windows also supports; the earlier
portability finding was specifically about test/$() shell syntax, not
&&). Replace scripts/build-check.mjs with a narrower
scripts/check-untracked-artifacts.mjs (from a new
src/scripts/check-untracked-artifacts.mts source) whose only job is
the git-ls-files-based untracked-file check. A tampered committed
check-untracked-artifacts.mjs is now caught by the outer, non-generated
git diff comparison, which rebuilds and re-diffs it before it ever
runs.

Verified locally: pnpm run build:check still passes on a clean tree,
still fails on tracked drift, and still fails on an untracked
scripts/*.mjs probe under git -c status.showUntrackedFiles=no.

Refs #1707
