+++ .gitattributes
@@ -18,9 +18,9 @@ scripts/authoring-label-guard.mjs linguist-generated=true
 scripts/autopilot-suitability.mjs linguist-generated=true
 scripts/branch-conflict-state.mjs linguist-generated=true
 scripts/branch-name.mjs linguist-generated=true
-scripts/build-check.mjs linguist-generated=true
 scripts/build-ts.mjs linguist-generated=true
 scripts/check-pnpm-boundary.mjs linguist-generated=true
+scripts/check-untracked-artifacts.mjs linguist-generated=true
 scripts/ci-wait-policy.mjs linguist-generated=true
 scripts/ci-wait-state.mjs linguist-generated=true
 scripts/claim-approval-gate.mjs linguist-generated=true
