+++ package.json
@@ -56,7 +56,7 @@
     "lint:precommit": "biome check && dprint check \"**/*.md\" && markdownlint-cli2 \"**/*.md\"",
     "typecheck": "tsc -p tsconfig.json",
     "build": "node scripts/build-ts.mjs",
-    "build:check": "node scripts/build-check.mjs",
+    "build:check": "pnpm run build && git diff HEAD --exit-code -- scripts bin .gitattributes && node scripts/check-untracked-artifacts.mjs",
     "test": "pnpm run test:scripts",
     "test:scripts": "node --test tests/*.test.mts",
     "fixtures:update": "node scripts/update-fixtures.mjs",
